+++ Foundation/NSArray/NSMutableArray.h
@@ -38,7 +38,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 -(void)exchangeobjectAtIndex:(NSUInteger)index withobjectAtIndex:(NSUInteger)other;
 
 -(void)sortUsingSelector:(SEL)selector;
--(void)sortUsingFunction:(int (*)(id, id, void *))compare context:(void *)context;
+-(void)sortUsingFunction:(NSInteger (*)(id, id, void *))compare context:(void *)context;
 
 -(void)sortUsingDescriptors:(NSArray *)descriptors;
 -(void)filterUsingPredicate:(NSPredicate *)predicate;
