+++ Foundation/NSDictionary/NSDictionary.m
@@ -259,7 +259,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
   }
   else {
    NSEnumerator *state=[self keyEnumerator];
-   NSUInteger      count=[self count];
+   int      count=[self count];
    id            key;
 
    [coder encodeValueOfObjCType:@encode(int) at:&count];
